refactor side menu component and add lazy loading for images
